pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-document-converter"
version = "0.1.0"
description = "A tool for converting documents to a format optimized for Retrieval-Augmented Generation (RAG) systems"
readme = "README.md"
requires-python = ">=3.11"
license = {file = "LICENSE"}
authors = [
    {name = "Rose Pace", email = "rose.a.pace@gmail.com"},
]
dependencies = [
    "pyyaml>=6.0",
    "markdown>=3.7",
    "ollama>=0.4.7",
    "click>=8.1.8",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.49.0"]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[project.scripts]
rag-converter = "rag_converter.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/rag_converter"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310"]

[tool.isort]
profile = "black"
line_length = 100
